test: Migrate resource, shamap Beast tests to GTest (#7133)
