+++ src/tests/libxrpl/CMakeLists.txt
@@ -29,6 +29,8 @@ set(test_modules
     basics
     crypto
     json
+    resource
+    shamap
     tx
     protocol_autogen
 )
