+++ src/tests/libxrpl/shamap/SHAMapSync.cpp
@@ -1,0 +1,170 @@
+#include <xrpl/basics/Blob.h>
+#include <xrpl/basics/SHAMapHash.h>
+#include <xrpl/basics/Slice.h>
+#include <xrpl/basics/base_uint.h>
+#include <xrpl/basics/random.h>
+#include <xrpl/beast/utility/Journal.h>
+#include <xrpl/beast/xor_shift_engine.h>
+#include <xrpl/protocol/Serializer.h>
+#include <xrpl/shamap/SHAMap.h>
+#include <xrpl/shamap/SHAMapItem.h>
+#include <xrpl/shamap/SHAMapMissingNode.h>
+#include <xrpl/shamap/SHAMapTreeNode.h>
+
+#include <boost/smart_ptr/intrusive_ptr.hpp>
+
+#include <gtest/gtest.h>
+#include <helpers/TestSink.h>
+#include <shamap/common.h>
+
+#include <chrono>
+#include <cstdint>
+#include <list>
+#include <utility>
+#include <vector>
+
+namespace xrpl::tests {
+
+class SHAMapSyncTest : public ::testing::Test
+{
+protected:
+    beast::Journal const j_{TestSink::instance()};
+    beast::xor_shift_engine eng_;
+
+    boost::intrusive_ptr<SHAMapItem>
+    makeRandomAS()
+    {
+        Serializer s;
+
+        for (int d = 0; d < 3; ++d)
+            s.add32(randInt<std::uint32_t>(eng_));
+        return makeShamapitem(s.getSHA512Half(), s.slice());
+    }
+
+    bool
+    confuseMap(SHAMap& map, int count)
+    {
+        // add a bunch of random states to a map, then remove them
+        // map should be the same
+        SHAMapHash const beforeHash = map.getHash();
+
+        std::list<uint256> items;
+
+        for (int i = 0; i < count; ++i)
+        {
+            auto item = makeRandomAS();
+            items.push_back(item->key());
+
+            if (!map.addItem(SHAMapNodeType::TnAccountState, item))
+            {
+                ADD_FAILURE() << "Unable to add item to map";
+                return false;
+            }
+        }
+
+        for (auto const& item : items)
+        {
+            if (!map.delItem(item))
+            {
+                ADD_FAILURE() << "Unable to remove item from map";
+                return false;
+            }
+        }
+
+        if (beforeHash != map.getHash())
+        {
+            ADD_FAILURE() << "Hashes do not match " << beforeHash << " " << map.getHash();
+            return false;
+        }
+
+        return true;
+    }
+};
+
+TEST_F(SHAMapSyncTest, sync)
+{
+    TestNodeFamily f{j_}, f2{j_};
+    SHAMap source{SHAMapType::FREE, f};
+    SHAMap destination{SHAMapType::FREE, f2};
+
+    int const items = 10000;
+    for (int i = 0; i < items; ++i)
+    {
+        source.addItem(SHAMapNodeType::TnAccountState, makeRandomAS());
+        if (i % 100 == 0)
+            source.invariants();
+    }
+
+    source.invariants();
+    ASSERT_TRUE(confuseMap(source, 500));
+    source.invariants();
+
+    source.setImmutable();
+
+    int count = 0;
+    source.visitLeaves([&count]([[maybe_unused]] auto const& item) { ++count; });
+    EXPECT_EQ(count, items);
+
+    std::vector<SHAMapMissingNode> missingNodes;
+    source.walkMap(missingNodes, 2048);
+    EXPECT_TRUE(missingNodes.empty());
+
+    destination.setSynching();
+
+    {
+        std::vector<std::pair<SHAMapNodeID, Blob>> a;
+
+        ASSERT_TRUE(source.getNodeFat(SHAMapNodeID(), a, randBool(eng_), randInt(eng_, 2)));
+
+        ASSERT_FALSE(a.empty()) << "NodeSize";
+
+        ASSERT_TRUE(
+            destination.addRootNode(source.getHash(), makeSlice(a[0].second), nullptr).isGood());
+    }
+
+    do
+    {
+        f.clock().advance(std::chrono::seconds(1));
+
+        // get the list of nodes we know we need
+        auto nodesMissing = destination.getMissingNodes(2048, nullptr);
+
+        if (nodesMissing.empty())
+            break;
+
+        // get as many nodes as possible based on this information
+        std::vector<std::pair<SHAMapNodeID, Blob>> b;
+
+        for (auto& it : nodesMissing)
+        {
+            // Keep failures fatal here because this loop is data-dependent.
+            // non-deterministic number of times and the number of tests run
+            // should be deterministic
+            if (!source.getNodeFat(it.first, b, randBool(eng_), randInt(eng_, 2)))
+                FAIL() << "Unable to fetch node";
+        }
+
+        // Keep failures fatal here because this loop is data-dependent.
+        // non-deterministic number of times and the number of tests run
+        // should be deterministic
+        if (b.empty())
+            FAIL() << "No nodes returned";
+
+        for (auto const& i : b)
+        {
+            // Keep failures fatal here because this loop is data-dependent.
+            // non-deterministic number of times and the number of tests run
+            // should be deterministic
+            if (!destination.addKnownNode(i.first, makeSlice(i.second), nullptr).isUseful())
+                FAIL() << "Known node was not useful";
+        }
+    } while (true);
+
+    destination.clearSynching();
+
+    EXPECT_TRUE(source.deepCompare(destination));
+
+    destination.invariants();
+}
+
+}  // namespace xrpl::tests
